added path to aster added scan path to aster added aster to main
